requirements.txt
google-genai
python-dotenv
pygame
pyttsx3
google-generativeai